fix : record 날짜 및 type 수정

+++ runwithme/src/main/java/com/runwithme/runwithme/domain/record/dto/CoordinateDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class CoordinateDto {
-    private double latitude;
-    private double longitude;
+    private int latitude;
+    private int longitude;
 }
